# 参考及文档说明/接口文档/智能工单系统 API.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能工单系统 API",
    "description": "智能工单管理系统接口文档",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:10001",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/user/{userId}": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "修改用户信息",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "用户管理"
        ],
        "summary": "删除用户",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/status": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "启停用账号",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/roles": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "分配角色",
        "operationId": "assignRoles",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/password": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "重置密码",
        "operationId": "resetPassword",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/department": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "调整所属部门",
        "operationId": "assignDepartment",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDepartmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/workorder/{id}/dispatch": {
      "post": {
        "tags": [
          "工单管理"
        ],
        "summary": "派单/转派",
        "operationId": "dispatch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/register": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "注册",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/login": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "登录",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/page": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "用户分页列表",
        "operationId": "page_1",
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "roleCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageResultUserVO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateUserDTO": {
        "type": "object",
        "description": "修改用户基本信息请求",
        "properties": {
          "realName": {
            "type": "string",
            "description": "真实姓名",
            "maxLength": 50,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "description": "联系电话;传空字符串表示清空,不传(null)表示不修改",
            "maxLength": 20,
            "minLength": 0
          },
          "departmentId": {
            "type": "integer",
            "format": "int64",
            "description": "所属部门ID,可空表示不调整"
          }
        },
        "required": [
          "realName"
        ]
      },
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UpdateUserStatusDTO": {
        "type": "object",
        "description": "修改账号状态请求",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "账号状态:0禁用 1启用",
            "maximum": 1,
            "minimum": 0
          }
        },
        "required": [
          "status"
        ]
      },
      "AssignRoleDTO": {
        "type": "object",
        "description": "分配角色请求",
        "properties": {
          "roleIds": {
            "type": "array",
            "description": "角色ID列表,至少一个",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "minItems": 1
          }
        },
        "required": [
          "roleIds"
        ]
      },
      "ResetPasswordDTO": {
        "type": "object",
        "description": "重置密码请求",
        "properties": {
          "newPassword": {
            "type": "string",
            "description": "新密码",
            "maxLength": 2147483647,
            "minLength": 6
          }
        },
        "required": [
          "newPassword"
        ]
      },
      "AssignDepartmentDTO": {
        "type": "object",
        "description": "调整所属部门请求",
        "properties": {
          "departmentId": {
            "type": "integer",
            "format": "int64",
            "description": "部门ID"
          }
        },
        "required": [
          "departmentId"
        ]
      },
...
      "DispatchDTO": {
        "type": "object",
        "description": "派单/转派请求参数",
        "properties": {
          "handlerId": {
            "type": "integer",
            "format": "int64",
            "description": "目标处理人用户ID"
          },
          "remark": {
            "type": "string",
            "description": "派单/转派说明,可空",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "handlerId"
        ]
      },
...
      "RegisterDTO": {
        "type": "object",
        "description": "注册请求参数",
        "properties": {
          "username": {
            "type": "string",
            "description": "登录账号",
            "maxLength": 50,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "description": "密码",
            "maxLength": 2147483647,
            "minLength": 6
          },
          "confirmPassword": {
            "type": "string",
            "description": "确认密码,需与密码一致",
            "minLength": 1
          },
          "realName": {
            "type": "string",
            "description": "真实姓名",
            "maxLength": 50,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "description": "联系电话,可空",
            "maxLength": 20,
            "minLength": 0
          },
          "departmentId": {
            "type": "integer",
            "format": "int64",
            "description": "所属部门ID,可空"
          }
        },
        "required": [
          "confirmPassword",
          "password",
          "realName",
          "username"
        ]
      },
      "LoginDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "账号",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "description": "密码",
            "minLength": 1
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PageResultUserVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserVO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ResultPageResultUserVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultUserVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UserVO": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "用户id"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "realName": {
            "type": "string",
            "description": "真实姓名"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64",
            "description": "所属部门id"
          },
          "departmentName": {
            "type": "string",
            "description": "所属部门名称"
          },
          "phone": {
            "type": "string",
            "description": "联系电话,未填写时为空"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "状态：1启用 0停用"
          },
          "roles": {
            "type": "array",
            "description": "角色标识列表,如 SUBMITTER",
            "items": {
              "type": "string"
            }
          },
          "perms": {
            "type": "array",
            "description": "权限码列表,如 workorder:review",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}